Add user and conversation types to DB schema (#154)
Just a small PR to add these two new tables.

I also changed some of the columns on existing tables from INTEGER->BIGINT, because without this change Diesel generates INTEGER as i32 instead of i64, overwriting the existing i64 rust types. I could have done this as a new migration but I don't think this will affect anyone at this early stage seeing as the sqlite types are dynamically enforced and the code-genned Rust comes out the same.

+++ xmtp/migrations/2023-06_05-134021_accounts/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS accounts (
     id INTEGER PRIMARY KEY NOT NULL,
-    created_at INTEGER NOT NULL,
+    created_at BIGINT NOT NULL,
     serialized_key BLOB NOT NULL
 );
